Polyhedron demo: merge Mesh_3 demo plugins

+++ Polyhedron/demo/Polyhedron/Scene_implicit_function_item.cpp
@@ -436,7 +436,7 @@ Scene_implicit_function_item::draw_edges(CGAL::Three::Viewer_interface* viewer) 
     frame_->getMatrix(d_mat);
     //Convert the GLdoubles matrices in GLfloats
     for (int i=0; i<16; ++i){
-        f_mat.data()[i] = GLfloat(d_mat[i]);
+        f_mat.data()[i] = double(d_mat[i]);
     }
     program->setUniformValue("f_matrix", f_mat);
     program->setAttributeValue("colors", QVector3D(0.6,0.6,0.6));
