+++ Polyhedron/demo/Polyhedron/Scene_polylines_item.cpp
@@ -26,7 +26,7 @@ public:
 };
 
 void
-Scene_polylines_item::create_Sphere(double R) const
+Scene_polylines_item::create_Sphere(float R) const
 {
   create_flat_and_wire_sphere(R, positions_spheres, normals_spheres, positions_wire_spheres);
 }
